deps/v8/test/webkit/fast/js/Promise-resolve-expected.txt

Test Promise resolution.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS thisInOnFulfilled is undefined
PASS result is "hello"
PASS successfullyParsed is true

TEST COMPLETE
